Refactor: Simplify main.dart and introduce CalendarGrid widget
This commit refactors the `main.dart` file by removing the default
counter application and replacing it with a new `Root` widget.
The `Root` widget now displays a `CalendarGrid` widget.

The `CalendarGrid` widget is introduced to display a scrollable grid
of months and days. Each day is a button that, when pressed, shows
a modal bottom sheet with the selected date.

Additionally, new enum files `month.dart` and `day.dart` have been
created to represent months and days of the week, respectively.
An empty `appScaffold.dart` widget file has also been added.